+++ src/Domain/Helper/LinkedSheetsTitle.php
@@ -9,22 +9,41 @@
 
 namespace Proximum\Vimeet\Domain\Helper;
 
+use Proximum\Vimeet\Application\View\Agenda\SheetMetView;
 use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Domain\Repository\Meeting\RequestRepositoryInterface;
 
 class LinkedSheetsTitle
 {
-    public static function getSheetTitleView(Sheet $sheet) :string
-    {
-        $linkedSheetTitles = [];
-        if ($sheet->hasLinkedSheets()) {
-            $linkedSheets = $sheet->getLinkedSheets();
-            foreach ($linkedSheets->getSheets() as $linkedSheet) {
-                $linkedSheetTitles[] = $linkedSheet->getTitle();
-            }
+    /**
+     * @var RequestRepositoryInterface
+     */
+    private $requestRepository;
 
-            return implode(' - ', $linkedSheetTitles);
+    public function __construct(RequestRepositoryInterface $requestRepository)
+    {
+        $this->requestRepository = $requestRepository;
+    }
+
+    /**
+     * @param Sheet $userSheet
+     * @param Sheet $sheetMet
+     *
+     * @return SheetMetView[]
+     */
+    public function getSheetMetViews(Sheet $userSheet, Sheet $sheetMet): array
+    {
+        if (!$sheetMet->hasLinkedSheets()) {
+            return [new SheetMetView($sheetMet->getTitle(), false)];
         }
 
-        return $sheet->getTitle();
+        $linkedSheetTitles = [];
+        $linkedSheets = $sheetMet->getLinkedSheets();
+        foreach ($linkedSheets->getSheets() as $otherSheet) {
+            $isHighLighted = $this->requestRepository->hasApprovedMeetingRequest($userSheet, $otherSheet);
+            $linkedSheetTitles[] = new SheetMetView($otherSheet->getTitle(), $isHighLighted);
+        }
+
+        return $linkedSheetTitles;
     }
 }