+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- VIMEET-1914 - Bold on sheet titles
 ### Added
 - VIMEET-1950 - In back-office, all participants are assigned to meeting if option is activated.
 - VIMEET-1931 - Unallow link if sheet has a scheduled meeting